DotSlash Windows Shim reference implementation
Summary:
This is the DotSlash Windows Shim in Rust using std. Even with all the
size optimization tricks, it's still over 100KB. The version we use
internally is written C using only Windows APIs. This makes it really
small but hard to compile. But we now have a no_std Rust version that
only uses Windows APIs that is 5KB. That version will be added next -
this diff is here to prove that the the tests work.

Unfortunately DotSlash can't be made into a workspace because
`panic = "abort"` can't be set just for a workspace member. I am not
thrilled about python tests, but those are what we have internally. I am
gonna try to see if I can make it into a cargo integration test. Not
sure how that's going to play out with such a snow flaky build as the
no_std one.

Reviewed By: dtolnay

Differential Revision: D53487093

fbshipit-source-id: 8ae7cd3615f48283aa4a5f1dc6d9d314686d48e7