do `chmod -w` after the temp directory is moved for an artifact
Summary:
On my macOS M1:

```
/tmp% mkdir foo
/tmp% chmod -w foo
/tmp% mv foo bar
mv: rename foo to bar: Permission denied
```

Though on CentOS:

```
/tmp$ mkdir foo
/tmp$ chmod -w foo
/tmp$ mv foo bar
```

It seems `mv` on Mac as an issue moving a directory without
the read bit but Linux does not. (I'm not sure what the equivalent
is on Windows, so I didn't test there.)

Regardless, we can do the `chmod -w` after the `mv` to make
things work.

Reviewed By: zertosh

Differential Revision: D53558587

fbshipit-source-id: f06165060f6a3110b1106dd6b254babce05c7406