create a new tempdir if a provider fails
Summary:
Previously, there was this comment in the code:

```
// If there is a failure, should we create a new temp_dir_to_mv?
```

This diff addresses that comment. Specifically , it moves the
creation of `temp_dir_to_mv` and `fetch_destination` inside
the `for provider_config in &artifact_entry.providers` loop
so they get re-created each time. Though note that `file_lock`
is still created only once, which happens before the loop.

Reviewed By: zertosh

Differential Revision: D53558613

fbshipit-source-id: cddada557601799cd552c269b4d173e2a3d1356f