Support .tar.xz archives
Some projecst are only available as .tar.xz archives.
For example, [Zig releases](https://ziglang.org/download/)
for Linux and macOS are only available as .tar.xz archives.

This PR adds support to dotslash for .tar.xz archives,
using the xz2 crate to provide the functionality.
The xz2 create wraps the liblzma library.

In this change, the library is linked statically,
so that the resulting binary does not depend on the
liblzma library being installed on the system.

**File size impact**:

```
❯ uname -s
Darwin

❯ git checkout main
❯ cargo build --release
❯ wc -c target/release/dotslash
  985208 target/release/dotslash

❯ git checkout tarxz
❯ cargo build --release
❯ wc -c target/release/dotslash
 1052776 target/release/dotslash
```

This increases the binary size by 67,568 bytes on macOS.
It comes down to 16,704 if we use dynamic linking.
For now, I've chosen to use static linking.

**Testing**:
We'll need to update https://github.com/zertosh/dotslash_fixtures to
generate a .tar.xz archive as well to be able to add a fixture for it.

Resolves #10