Support .zip archives
Simillarly to #12, this adds support for .zip archives.

The code differentiates between the decompression step
(e.g. xz gzip, etc.) and the archive format (e.g. tar, zip).
So unlike .tar.gz, this adds a new ArchiveFormat.

In match arms, this adds the possibility of attempting to handle
.zip.gz or similar, but that's not allowed at the entry point,
so the code errors or panics if it encounters that.

To keep the file size down, the default features of the zip crate
were disabled.
Namely, support for encrypted archives and bzip2 was left out.
There's also a 'time' feature flag that enables integration
with the [time crate](https://github.com/time-rs/time)
that was also left out.

**File size impact:**

```
❯ git checkout main
❯ cargo build --release
❯ wc -c target/release/dotslash
  985208 target/release/dotslash

❯ git checkout zip
❯ cargo build --release
❯ wc -c target/release/dotslash
 1018312 target/release/dotslash
```

Currently, this change adds 33,104 bytes to the binary size on macOS.
Enabling the bzip2 feature flag on zip will add another 33,232 bytes.

**License:** zip-rs is licensed under MIT, so it's safe to use here.

**Testing**:
Besides accompanying unit tests,
tested manually with the following file per #8:

```

{
  "name": "dprint",
  "platforms": {
    "macos-aarch64": {
      "size": 7030235,
      "hash": "blake3",
      "digest": "a16c60d1692ad21e12f9d7b167b3dc9d4e5f34c2c1afd5e705842f5ab3e94e0e",
      "format": "zip",
      "path": "dprint",
      "providers": [
        {
          "url": "https://github.com/dprint/dprint/releases/download/0.45.0/dprint-aarch64-apple-darwin.zip"
        }
      ]
    }
  }
}
```

Resolves #8