add a CHANGELOG.md
Summary:
I am open to suggestions on how we want to format this,
but it seems nice to have the various links handy and the
date alongside the version so you don't have to go and
hunt for it.

Also, while I know the GitHub URLs seem a bit long and
unwieldy, I believe they get shortened quite a bit when
the file is rendered on GitHub, which is how I expect most
users will consume it.

Reviewed By: zertosh

Differential Revision: D53780805

fbshipit-source-id: f35f2f41851fd48181280997bf847287c3ad5a39