update website to include logo/favicon (#29)
Summary:
This commit makes use of the SVG assets that were sitting in `assets/Final/SVG` and features them on the website so there is a proper logo and favicon. Here is how things look in dark and light themes.

I'm not particularly attached to any of the specifics of this design, so the maintainers/reviewers should feel free to change it however they see fit. My primary motivation is having proper branding on the site (particularly the favicon so I can find the page easily among all my Chrome tabs!).

![Screenshot from 2024-05-27 11-08-11](https://github.com/facebook/dotslash/assets/655869/cada26e4-01d9-444a-a036-0eba40e793db)

![Screenshot from 2024-05-27 11-09-20](https://github.com/facebook/dotslash/assets/655869/543e32de-640c-4836-85cb-0acae1453755)

![Screenshot from 2024-05-27 11-09-30](https://github.com/facebook/dotslash/assets/655869/91e250f3-698c-400e-902c-0056da7e23e2)

![Screenshot from 2024-05-27 11-09-37](https://github.com/facebook/dotslash/assets/655869/f7b2b2e2-d94b-4713-946f-7d3903f2cdc1)

While here, I also cleared out a number of stock images from the original Docusaurus template in `static/img` since they are unnecessary / could create confusion.

Pull Request resolved: https://github.com/facebook/dotslash/pull/29

Reviewed By: dtolnay

Differential Revision: D57847383

Pulled By: zertosh

fbshipit-source-id: 2dd44fa7f63f54729aa4150aa859d66c4d8ab042