Release statically linked executables on Linux (#34)
Summary:
One of the major benefits of dotslash is to be used to bootstrap other tools on pretty much any system. Unfortunately when it comes to linux the executables included in the release are built on Ubuntu 22.04 which means that they link a fairly modern glibc. Trying to use the executable on an older Ubuntu (or any other flavor of Linux that doesn't use glibc) will fail due to linking errors.

We could instead build the binary on a super old linux version but Github only goes back to 20.04 so instead let's produce a fully static executable that statically links the musl libc.

This is a common pattern that many tools built for linux adopt and it will greatly simplify the adoption on linux distros that don't repackage dotslash.

Pull Request resolved: https://github.com/facebook/dotslash/pull/34

Reviewed By: dtolnay

Differential Revision: D64218797

Pulled By: zertosh

fbshipit-source-id: 308cc09fece777801c0cae3c15f19c0ffaf333fc