Replace "decompress" with more general "unarchive"
Summary:
`unpack_file` with `ArchiveType` is less awkward than `untar` with a
boolean "is_tar_gz". This also sets up being able to use zstd in the
internal dotslash.

`flate2::read::GzDecoder` is switched to `flate2::bufread::GzDecoder`,
`xz2::read::XzDecoder` is switched to `xz2::bufread::XzDecoder`, and
`ZstdDecoder::new` is switched to `ZstdDecoder::with_buffer` to avoid
double wrapping the reader with `BufReader`. The new calls make it
explicit that a `BufReader` must be passed (because the internal
decompressor won't do it for you).

Reviewed By: diliop

Differential Revision: D65046088

fbshipit-source-id: 63bdea2e3bf499a2086f330c29e27b01838546c0