Update syn from 2.0.85 to 2.0.89
Summary:
Required by new versions of `thiserror`. This diff is split out from the thiserror v2 update diff so that all the things that depend on syn *not* via thiserror do not need to run CI on that diff, making it more likely that I can identify all the places that need code changes going from thiserror v1 to v2.

Release notes:

- https://github.com/dtolnay/syn/releases/tag/2.0.86
- https://github.com/dtolnay/syn/releases/tag/2.0.87
- https://github.com/dtolnay/syn/releases/tag/2.0.88
- https://github.com/dtolnay/syn/releases/tag/2.0.89

Reviewed By: zertosh

Differential Revision: D66555204

fbshipit-source-id: ea843cfae0f918292ca251c70beac6578542ed76