Update cc version 1.0.90 -> 1.2.10
Summary:
Update cc, we need an updated version so we can update other crates we care about.

For the two crates that insisted on the old version I created a fork from the same version we're using and changed Cargo.toml in the fork. I couldn't find a way to patch it in place. Anyway that's a temporary hack as hopefully I'll eventually succeed in updating tree-sitter.

Reviewed By: jsgf

Differential Revision: D68669984

fbshipit-source-id: b7fea9136c1d24c8158f0366012c036e98d9e33e