[RFC] config: Support provider randomization
**Proposed change**

This changes the dotslash file schema to add a new optional field:

```diff
 {
     // ...
     providers: [ ... ]
+    providers_order: "sequential" | "random",
 }
```

The default ("sequential") will try the providers in the order
they are specified in the 'providers' list.
This is how dotslash behaves today.

If the value is "random",
the order in which providers are tried will be randomized.

**Why**

This will result in better usage patterns
when there are multiple providers for an executable.
As an example, for the [Zig compiler](https://ziglang.org/),
while there's an officially hosted source,
it's preferred to use third-party mirrors
(https://github.com/mlugg/setup-zig/blob/153c8d5202cbb8c7e10831110a3afd27593eb960/mirrors.json)
to avoid hammering the official server.

**Binary size**

This adds ~16 KB to the output of `cargo build --release`:

    Before  1103376 bytes
    After   1119904 bytes
    ------- -------------
    Change    16528 bytes

Resolves #33