config: Support provider randomization (#49)
Summary:
**Proposed change**

This changes the dotslash file schema to add a new optional field:

```diff
 {
     // ...
     providers: [ ... ]
+    providers_order: "sequential" | "weighted-random",
 }
```

The default ("sequential") will try the providers in the order
they are specified in the 'providers' list.
This is how dotslash behaves today.

If the value is "weighted-random",
the order in which providers are tried will be randomized.

By default, all providers will be weighted equally.
This can be changed by adding a `"weight"` value
to the provider configuration:

```json
    {"url": "https://primary.example.com/hermes.tar.gz", "weight": 3},
    {"url": "https://mirror1.example.com/hermes.tar.gz", "weight": 1},
    {"url": "https://mirror2.example.com/hermes.tar.gz", "weight": 1},
    {"url": "https://mirror3.example.com/hermes.tar.gz", "weight": 1}
```

Weight must be an integer >=1.
Zero ("don't select") is not permitted.

To shuffle providers in weighted order, they are sampled without replacement.
That is, even if the weight of a provider would present it twice before another option,
it is not re-added to the final order of providers—it is tried only once.

**Why**

This will result in better usage patterns
when there are multiple providers for an executable.
As an example, for the [Zig compiler](https://ziglang.org/),
while there's an officially hosted source,
it's preferred to use third-party mirrors
(https://github.com/mlugg/setup-zig/blob/153c8d5202cbb8c7e10831110a3afd27593eb960/mirrors.json)
to avoid hammering the official server.

**Binary size**

This adds ~16 KB to the output of `cargo build --release` on macOS ARM64:

    Before  1103440 bytes
    After   1120000 bytes
    ------- -------------
    Change    16560 bytes

Resolves https://github.com/facebook/dotslash/issues/33

Pull Request resolved: https://github.com/facebook/dotslash/pull/49

Reviewed By: dtolnay

Differential Revision: D75089952

Pulled By: bigfootjon

fbshipit-source-id: 2e8af277d564fc0291c67a75a3214bb735cdf270