Enable getrandom js feature
Summary: Required to enable Pyrefly Wasm to build with the Cargo.lock we have, meaning we can build a `Cargo.lock` for all of Pyrefly. That requires the `js` feature of `getrandom` (`getrandom` seems to be the only library that works like this). No direct dependencies, but loads of transitive dependencies. I pinned to 0.2.16 as that is the latest release, which has a minor backport, as it is the newest 0.2 version.

Reviewed By: kinto0

Differential Revision: D76505161

fbshipit-source-id: 04d1bb8b16ba38fd359e5f2c52f7feb1393b2769