fix: upload non-version tagged release artifacts
Our Dev Container feature, today, uses the GitHub REST API in order to
figure out the correct "latest" release.  However, this can be rate
limited, and in fact, we've had some of our tests fail for that exact
reason.  We cannot use the magic `latest` URL to fetch artifacts because
the version number is also in the artifact.  Therefore, we must start to
upload artifacts without version numbers as well.

This also makes a change with the non-versioned artifacts to use
`uname -m`-compatible architecture suffixes.