fix: upload non-version tagged release artifacts (#64)
Summary:
Our Dev Container feature, today, uses the GitHub REST API in order to figure out the correct "latest" release.  However, this can be rate limited, and in fact, we've had some of our tests fail for that exact reason.  We cannot use the magic `latest` URL to fetch artifacts because the version number is also in the artifact.  Therefore, we must start to upload artifacts without version numbers as well.

This also makes a change with the non-versioned artifacts to use `uname -m`-compatible architecture suffixes.

Pull Request resolved: https://github.com/facebook/dotslash/pull/64

Reviewed By: bigfootjon

Differential Revision: D77187824

Pulled By: sdwilsh

fbshipit-source-id: 8de4eb0eeb3f73bc77a0d8bd7f7812a950931852