feat: Create a devcontainer (#65)
Summary:
Currently, this just supports the development of the feature, but I did set it up to be based off of the Rust image, so it should be possible for us to use it for building, tests, and clippy.

This will help iterate on improvements to the feature.

Pull Request resolved: https://github.com/facebook/dotslash/pull/65

Reviewed By: bigfootjon

Differential Revision: D77247747

Pulled By: sdwilsh

fbshipit-source-id: cc7ea9f4ea68098ff4d007936fae63d2f065edbd