feat: allow one-line install of the latest version (#72)
Summary:
This re-implements https://github.com/facebook/dotslash/issues/17 now that we have fixed the underlying issues that caused it to get reverted in the first place.

I've done the documentation a bit differently to try and make it a bit cleaner.  I also did not update the versioned artifacts that are due for removal in https://github.com/facebook/dotslash/issues/68.

Pull Request resolved: https://github.com/facebook/dotslash/pull/72

Reviewed By: bigfootjon

Differential Revision: D77397668

Pulled By: sdwilsh

fbshipit-source-id: fdf9d7b90bfade637ed2acf1934460c5d0739abf