Point internal users to the Wiki
Summary:
A number of folks have/tried to install the public/OSS version of
DotSlash on their machines, which doesn't work with internal use cases. A hunch
is that an LLM might be pointing them to the public docs:
https://www.internalfb.com/intern/staticdocs/dotslash/docs/installation/.

This diff adds an explicit note for internal employees to not install the
public/OSS version. Instead, it points them to an Internal Wiki for
troubleshooting.

Reviewed By: bigfootjon

Differential Revision: D81272708

fbshipit-source-id: 779c46d53bc607ce4a4544d83dd8c198b33fd641