Create an S3 provider for fetching artifacts
This allows you to fetch artifacts from an S3 bucket which is not
publicly accessible. You can combine this with the HTTP (default)
provider to try to fetch via HTTP and fallback to using an authenticate
client so that fetching will work in many environments. Note, using this
provider requires some sort of credential source, such as aws-vault.