Create an S3 provider for fetching artifacts (#79)
Summary:
We've been rolling out usage of dotslash at Stile, thanks for building such a useful tool! One thing that has been useful for us is to be able to download binaries from S3 when you don't necessarily have HTTP access (e.g. private bucket, HTTP access locked down to particular IPs, etc). I've been trialing the S3 provider for a little bit so am hoping to upstream it. There seems to be some wider interest in it based on https://github.com/facebook/dotslash/issues/30.

I've followed the same style as the GitHub and HTTP providers of using a CLI. So, this requires that the `aws` CLI is installed.

Closes: https://github.com/facebook/dotslash/issues/30

Pull Request resolved: https://github.com/facebook/dotslash/pull/79

Reviewed By: dtolnay

Differential Revision: D82117024

Pulled By: bigfootjon

fbshipit-source-id: 482736b76da3725969b1ead0bb10b9c8c7d15bc0