Enable windows-sys Win32_Security_Credentials feature
Summary:
This feature is required in order for the `windows_sys` crate to contain `WIn32::Security::Credentials`, which is used here:

https://www.internalfb.com/code/fbsource/[84251e005f202c8ab26688bd56f5b707c12b2e03]/fbcode/tools/px/keyring.rs?lines=57-63

Previously the feature was enabled by a transitive dependency through some other crate we depend on, but in D87988428 that transitive dependency went away and the feature got turned off. Since our first-party code requires the feature, it should be listed in our Cargo.toml.

Reviewed By: cfsmp3

Differential Revision: D87991526

fbshipit-source-id: 9c99857171e96d6cf80f9e5dc4484945886a3a5b